+++ src/flockwave/server/ext/mavlink/enums.py
@@ -1,9 +1,28 @@
-from enum import IntEnum, IntFlag
+from enum import Enum, IntEnum, IntFlag
 from flockwave.server.model.gps import GPSFixType as OurGPSFixType
 from struct import Struct
 from typing import Union
 
 __all__ = ("MAVComponent",)
+
+
+class ConnectionState(Enum):
+    """Enum representing the possible connection states between the GCS and a
+    drone.
+    """
+
+    DISCONNECTED = "disconnected"
+    """Drone is disconnected, we have not received a heartbeat from it
+    recently.
+    """
+
+    SLEEPING = "sleeping"
+    """Drone is in a sleep state; we are receiving heartbeats but they indicate
+    that the flight controller is not fully powered on.
+    """
+
+    CONNECTED = "connected"
+    """Drone is fully connected."""
 
 
 class MAVAutopilot(IntEnum):