+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Fixed a bug with drones that are capable of entering a sleep state where the
+  autopilot type was not detected correctly after the drone was woken up.
+
 ## [2.16.0] - 2023-10-25
 
 ### Added